// examples/secure-api/generated/notes.gen.ts
export const CreateNoteBodySchema = z.object({
  title: z.string(),
  body: z.string().optional(),
}).strict()
export type CreateNoteBody = z.infer<typeof CreateNoteBodySchema>

export const NoteSchema = z.object({
  id: z.string(),
  ownerId: z.string(),
  title: z.string(),
  body: z.string().optional(),
  createdAt: z.string().datetime(),
}).strict()
export type Note = z.infer<typeof NoteSchema>

export const CreateNoteOutputSchema = z.object({
  item: NoteSchema,
}).strict()
export type CreateNoteOutput = z.infer<typeof CreateNoteOutputSchema>

export const GetNoteOutputSchema = z.object({
  item: NoteSchema,
}).strict()
export type GetNoteOutput = z.infer<typeof GetNoteOutputSchema>

export const ListNotesOutputSchema = z.object({
  items: z.array(NoteSchema),
  nextToken: z.string().optional(),
}).strict()
export type ListNotesOutput = z.infer<typeof ListNotesOutputSchema>